requirements.txt
grpcio==1.64.1
grpcio-tools==1.64.1
protobuf==5.27.1
PyYAML==6.0.1
tabulate==0.9.0
